modify runtime code
